# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strategy Engine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateClientInput {
  name: string;
  websiteUrl: string;
  instagramHandle: string;
  oneLineDescription: string;
}

export type GenerateStrategyInputTemplateType =
  (typeof GenerateStrategyInputTemplateType)[keyof typeof GenerateStrategyInputTemplateType];

export const GenerateStrategyInputTemplateType = {
  performance_marketing: "performance_marketing",
  brand_building: "brand_building",
  personal_brand: "personal_brand",
  d2c_growth: "d2c_growth",
  auto: "auto",
} as const;

export interface GenerateStrategyInput {
  templateType?: GenerateStrategyInputTemplateType;
}

export type UpdateStrategyInputStructuredStrategy = { [key: string]: unknown };

export type UpdateStrategyInputStatus =
  (typeof UpdateStrategyInputStatus)[keyof typeof UpdateStrategyInputStatus];

export const UpdateStrategyInputStatus = {
  draft: "draft",
  approved: "approved",
} as const;

export interface UpdateStrategyInput {
  strategyDocument?: string;
  structuredStrategy?: UpdateStrategyInputStructuredStrategy;
  status?: UpdateStrategyInputStatus;
}

export interface Client {
  id: string;
  name: string;
  website?: string | null;
  instagramHandle?: string | null;
  oneLineDescription?: string | null;
  createdAt: string;
}

export type ClientSummary = Client & {
  hasStrategy?: boolean;
  templateType?: string | null;
  status?: string | null;
};

export type EnrichedDataTargetAudience = {
  who?: string;
  age_range?: string;
  stage?: string;
};

export interface EnrichedData {
  brand_name?: string;
  offer?: string;
  price_range?: string;
  business_model?: string;
  target_audience?: EnrichedDataTargetAudience;
  brand_tone?: string;
  positioning?: string;
  platform?: string;
  content_preference?: string;
  competitors?: string[];
}

export type OnboardingProfileRawInput = { [key: string]: unknown };

export interface OnboardingProfile {
  id: string;
  clientId: string;
  rawInput: OnboardingProfileRawInput;
  enrichedData?: EnrichedData | null;
}

export type StrategyStructuredStrategy = { [key: string]: unknown };

export interface Strategy {
  id: string;
  clientId: string;
  structuredStrategy: StrategyStructuredStrategy;
  strategyDocument: string;
  templateType: string;
  version: number;
  status: string;
  createdAt?: string;
  updatedAt?: string;
}

export interface ClientDetail {
  client: Client;
  onboarding?: OnboardingProfile | null;
  strategy?: Strategy | null;
}