# lib/api-zod/src/generated/types/client.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strategy Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface Client {
  id: string;
  name: string;
  website?: string | null;
  instagramHandle?: string | null;
  oneLineDescription?: string | null;
  createdAt: string;
}