# lib/api-zod/src/generated/types/clientDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strategy Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { Client } from "./client";
import type { OnboardingProfile } from "./onboardingProfile";
import type { Strategy } from "./strategy";

export interface ClientDetail {
  client: Client;
  onboarding?: OnboardingProfile | null;
  strategy?: Strategy | null;
}