# lib/api-zod/src/generated/types/clientSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strategy Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { Client } from "./client";

export type ClientSummary = Client & {
  hasStrategy?: boolean;
  templateType?: string | null;
  status?: string | null;
};