# lib/api-zod/src/generated/types/createClientInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strategy Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateClientInput {
  name: string;
  websiteUrl: string;
  instagramHandle: string;
  oneLineDescription: string;
}