# lib/api-zod/src/generated/types/enrichedData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strategy Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { EnrichedDataTargetAudience } from "./enrichedDataTargetAudience";

export interface EnrichedData {
  brand_name?: string;
  offer?: string;
  price_range?: string;
  business_model?: string;
  target_audience?: EnrichedDataTargetAudience;
  brand_tone?: string;
  positioning?: string;
  platform?: string;
  content_preference?: string;
  competitors?: string[];
}