# lib/api-zod/src/generated/types/generateStrategyInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strategy Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateStrategyInputTemplateType } from "./generateStrategyInputTemplateType";

export interface GenerateStrategyInput {
  templateType?: GenerateStrategyInputTemplateType;
}