# lib/api-zod/src/generated/types/onboardingProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strategy Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { EnrichedData } from "./enrichedData";
import type { OnboardingProfileRawInput } from "./onboardingProfileRawInput";

export interface OnboardingProfile {
  id: string;
  clientId: string;
  rawInput: OnboardingProfileRawInput;
  enrichedData?: EnrichedData | null;
}