# lib/api-zod/src/generated/types/strategy.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strategy Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { StrategyStructuredStrategy } from "./strategyStructuredStrategy";

export interface Strategy {
  id: string;
  clientId: string;
  structuredStrategy: StrategyStructuredStrategy;
  strategyDocument: string;
  templateType: string;
  version: number;
  status: string;
  createdAt?: string;
  updatedAt?: string;
}