# lib/api-zod/src/generated/types/updateStrategyInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strategy Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateStrategyInputStatus } from "./updateStrategyInputStatus";
import type { UpdateStrategyInputStructuredStrategy } from "./updateStrategyInputStructuredStrategy";

export interface UpdateStrategyInput {
  strategyDocument?: string;
  structuredStrategy?: UpdateStrategyInputStructuredStrategy;
  status?: UpdateStrategyInputStatus;
}